Moving chat rooms loading logic to component
